# docs/swagger.yaml
basePath: /
definitions:
  models.Subscription:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      id:
        type: integer
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a sample server for GSES BTC Application.
  title: GSES BTC Application API
  version: "1.0"
paths:
  /api/rate:
    get:
      description: Get the current exchange rate of USD to UAH using a third-party
        API
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: number
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get current USD to UAH exchange rate
      tags:
      - rate
  /api/subscribe:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Subscribe an email to receive daily updates on the USD to UAH exchange
        rate
      parameters:
      - description: Email Address
        in: formData
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Subscribe to daily exchange rate updates
      tags:
      - subscription
  /api/subscriptions:
    get:
      description: Get all email subscriptions from the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Subscription'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all subscriptions
      tags:
      - subscription
swagger: "2.0"