# models/subscription.go
package models

import (
	"time"
)

type Subscription struct {
	ID        uint       `gorm:"primary_key" json:"id"`
	CreatedAt time.Time  `json:"created_at"`
	UpdatedAt time.Time  `json:"updated_at"`
	DeletedAt *time.Time `sql:"index" json:"deleted_at,omitempty"`
	Email     string     `gorm:"unique;not null" json:"email"`
}